Added some wheel management features in it's own deticated menu.
Added an option to export only selected wheels to a file using a multi-selection list in the wheel manager menu.
Made it so the app checks for duplicate wheels before Proceeding to import them.
Added confirmation prompts for exporting and importing all existing wheels from a file.
Added the ability to export and import all existing wheels from a file.
Added the ability to spin a random wheel from the ones that have been created by you.