Nearly all of the sections of the app are now translatable.
The spin settings section of the app is now translatable.
The second spin wheel section of the app is now translatable.
The edit wheel section of the app is now translatable.
The manage wheel section of the app is now translatable.
The add wheel section of the app is now translatable.